% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{available_modules}
\alias{available_modules}
\title{Available modules}
\usage{
available_modules(map)
}
\arguments{
\item{map}{Name of a \code{grr} map generated by \code{create_from_template()}.}
}
\value{
A S3 object of class \code{grr} built upon a list.
}
\description{
Lists the modules that are available in a \code{grr} map made from a template.
}
