% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\docType{data}
\name{class_commons}
\alias{class_commons}
\title{Class commons}
\format{
A list with a bunch of stuff in it:
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
class_commons
}
\description{
A classroom space to engage with course content in a social manner. Designed
for roughly 24 students plus one instructor.
}
\keyword{datasets}
