% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_from_template.R
\name{create_from_template}
\alias{create_from_template}
\title{Create new \code{gatherer} map from template}
\usage{
create_from_template(template, remove_modules = "", ...)
}
\arguments{
\item{template}{Name of template. Run \code{list_templates} to see current options}

\item{remove_modules}{A character vector indicating modules from the template
that will be removed. See template documentation with \code{vignette("template")}
for options.}

\item{...}{Additional arguments passed to}
}
\value{
A S3 object of class \code{gatherer} built upon a list.
}
\description{

}
