% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_template}
\alias{create_template}
\title{Create modules for a template}
\usage{
create_template(obj, title, description, ...)
}
\arguments{
\item{...}{One or more modules as lists, each separated by a comma.}
}
\value{
A S3 object of class \code{grr_mod} built upon a list.
}
\description{
Creates a list of \code{grr_mod} modules, each containing at least one
\code{grr_obj}.
}
