% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map.R
\name{get_map}
\alias{get_map}
\title{Get a map from gather.town}
\usage{
get_map(api_key, space_id, map_id)
}
\arguments{
\item{api_key}{Your API key from gather.town, passed as a string.}

\item{space_id}{Your space ID, which is the string of random
characters and space name from of your space's URL; the portion in
of the URL that follows \verb{https://gather.town/app}.}

\item{map_id}{Your map ID as a string, which is the same name as the
map file that you would load at \url{https://gather.town/mapmaker}.}
}
\value{
A named list containing all of the information that defines
the map
}
\description{
This function fetches your map from gather.town's API and makes it
available as a named list.
}
