% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{load_grr}
\alias{load_grr}
\title{Load \code{grr} object from a template}
\usage{
load_grr(template, modules = "all")
}
\arguments{
\item{template}{Name of template. Run \code{list_templates()} to see current options}

\item{modules}{A character vector indicating modules from the template
to include.}
}
\value{
A S3 object of class \code{grr} built upon a list.
}
\description{
Loads a new \code{grr} map from one of several template options.
}
