% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatherer_class.R
\name{new_gatherer}
\alias{new_gatherer}
\title{Construct a new \code{gatherer} object}
\usage{
new_gatherer(
  map_file = list(),
  map_id = character(),
  space_id = character(),
  api_key = character()
)
}
\arguments{
\item{map_file}{A named list corresponding to the json file that's
readable by gather.town.}

\item{template_info}{A named list of information corresponding to the
template from which the object was created.}
}
\value{
A S3 object of class \code{gatherer} built upon a list.
}
\description{
A constructor to create a new \code{gatherer} S3 object.
}
