% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grr.R
\name{new_grr}
\alias{new_grr}
\title{Construct a new \code{grr} object}
\usage{
new_grr(
  map_file = list(),
  map_id = character(),
  space_id = character(),
  api_key = character()
)
}
\arguments{
\item{map_file}{A named list corresponding to the json file that's
readable by gather.town.}

\item{template_info}{A named list of information corresponding to the
template from which the object was created.}
}
\value{
A S3 object of class \code{grr} built upon a list.
}
\description{
A constructor to create a new \code{grr} S3 object with a blank template.
}
