% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grr.R
\name{new_grr_mod}
\alias{new_grr_mod}
\title{Construct a new \code{grr_mod} object}
\usage{
new_grr_mod(module)
}
\arguments{
\item{map_file}{A named list corresponding to the json file that's
readable by gather.town.}

\item{template_info}{A named list of information corresponding to the
template from which the object was created.}
}
\value{
A S3 object of class \code{grr_mod} built upon a list.
}
\description{
A constructor to create a new \code{grr_mod} S3 object.
}
