% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatherer_class.R
\name{plot.gatherer}
\alias{plot.gatherer}
\title{Plot a \code{gatherer} background map}
\usage{
\method{plot}{gatherer}(x, y, ...)
}
\arguments{
\item{map}{A map object of class \code{gatherer}}
}
\description{
Generates a raster image of the png background map.
}
