% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grr.R
\name{plot.grr}
\alias{plot.grr}
\title{Plot a background map}
\usage{
\method{plot}{grr}(x, y, ...)
}
\arguments{
\item{map}{A map object of class \code{grr}}
}
\description{
Generates a raster image of the png background map.
}
