% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_api.R
\name{post_map}
\alias{post_map}
\title{Post a map on gather.town}
\usage{
post_map(api_key, space_id, map_id, map_file, ...)
}
\arguments{
\item{api_key}{An API key from gather.town, passed as a string.}

\item{space_id}{A space ID, which is the string of random
characters and space name from of your space's URL; the portion in
of the URL that follows \verb{https://gather.town/app}.}

\item{map_id}{A map ID as a string, which is the same name as the
map file that would load at \url{https://gather.town/mapmaker}.}

\item{map_file}{The map file, as named list.}
}
\description{
Pushes a map file from R and posts it on gather.town.
}
