% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_api.R
\name{pull_map}
\alias{pull_map}
\title{Pull the map file from gather.town}
\usage{
pull_map(map)
}
\arguments{
\item{map}{An object of class \code{"gatherer"}.}
}
\value{
An updated object of class \code{"gatherer"}
}
\description{
Pulls the map file from gather.town and merges it into the map
object that is passed to the function.
}
