% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_api.R
\name{push_map}
\alias{push_map}
\title{Push a map to gather.town}
\usage{
push_map(map)
}
\arguments{
\item{map}{An object of class \code{"gather_map"}.}
}
\description{
Pushes a map file from R and posts it on gather.town.
}
