% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grr.R
\name{summary.grr}
\alias{summary.grr}
\title{Summary of a \code{grr} map}
\usage{
\method{summary}{grr}(obj, ...)
}
\arguments{
\item{map}{A map object of class \code{grr}.}
}
\description{
Prints useful summary information about a \code{grr} object.
}
